fix: move access hooks next to their consumers
Move useTrustySettingsAccessAllowed into content/ (next to
useTrustyCRState) and useNIMSettingsAccessAllowed into
pages/projectSettings/ (next to NIMSettingsCard). Update all imports.

RHOAIENG-63340
