+++ frontend/src/pages/projects/projectSettings/__tests__/ModelBiasSettingsCard.spec.tsx
@@ -2,12 +2,12 @@ import * as React from 'react';
 import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import '@testing-library/jest-dom';
-import { useTrustySettingsAccessAllowed } from '#~/concepts/trustyai/useTrustySettingsAccessAllowed';
+import { useTrustySettingsAccessAllowed } from '#~/concepts/trustyai/content/useTrustySettingsAccessAllowed';
 import useManageTrustyAICR from '#~/concepts/trustyai/useManageTrustyAICR';
 import { TrustyInstallState } from '#~/concepts/trustyai/types';
 import ModelBiasSettingsCard from '#~/pages/projects/projectSettings/ModelBiasSettingsCard';
 
-jest.mock('#~/concepts/trustyai/useTrustySettingsAccessAllowed');
+jest.mock('#~/concepts/trustyai/content/useTrustySettingsAccessAllowed');
 jest.mock('#~/concepts/trustyai/useManageTrustyAICR');
 
 const mockUseTrustySettingsAccessAllowed = jest.mocked(useTrustySettingsAccessAllowed);
