+++ packages/nim-serving/src/pages/projectSettings/__tests__/NIMSettingsCard.spec.tsx
@@ -2,11 +2,11 @@ import * as React from 'react';
 import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import '@testing-library/jest-dom';
-import { useNIMSettingsAccessAllowed } from '../../../api/accounts/useNIMSettingsAccessAllowed';
+import { useNIMSettingsAccessAllowed } from '../useNIMSettingsAccessAllowed';
 import useNIMAccountStatus, { NIMAccountStatus } from '../../../api/accounts/hooks';
 import NIMSettingsCard from '../NIMSettingsCard';
 
-jest.mock('../../../api/accounts/useNIMSettingsAccessAllowed');
+jest.mock('../useNIMSettingsAccessAllowed');
 
 jest.mock('../../../api/accounts/hooks', () => ({
   __esModule: true,
